requirements.txt
pyarrow==15.0.0
pytest==7.4.4
pandas==2.2.0
Flask==3.0.1
Flask-Cors==4.0.0
pyodbc==5.0.1
azure-storage-blob==12.19.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pypdf==4.0.0
openai==1.11.0
langchain-openai==0.0.5
chromadb==0.4.22
jupyter==1.0.0
nltk==3.8.1
langchain==0.1.3
langchain-community==0.0.16
sentence-transformers==2.3.1
pinecone-client==3.0.2
tiktoken==0.5.2
mysql-connector==2.2.9
mysql-connector-python==2.2.9
